requirements.txt
matplotlib
numpy
opencv_python
Pillow
scipy
scikit-image
tqdm
efficientnet_pytorch
torch
torchvision
adjustText
grad_cam
ipython
pandas
scikit_learn
seaborn
statsmodels
umap_learn
utils
spyder
jupyterlab
split-folders
wandb
giddy
networkx
scanpy
phate
scvelo
